// payload-types.ts
export interface User {
  id: string;
  name?: string | null;
  username?: string | null;
  bio?: string | null;
  avatar?: string | Media | null;
  authProvider: "local" | "google" | "facebook" | "x";
  sub?: string | null;
  roles?: ("admin" | "user")[] | null;
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  password: string | null;
}
// ...
export interface UserPoint {
  id: string;
  user: string | User;
  points: number;
  extraPoints?: number;
  pointType?: "PLUS" | "PAYG" | "FREE";
  expireAt?: string | null;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface Media {
  id: string;
  alt: string;
  caption?:
    | {
        [k: string]: unknown;
      }[]
    | null;
  prefix?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
}